Keeping old lib as default
