+++ Cargo.toml
@@ -19,6 +19,6 @@ rp2040-hal = { version = "0.11.0", optional = true }
 rp235x-hal = { version = "0.2.0", optional = true }
 
 [features]
-default = []
+default = ["rp2040"]
 rp2040 = ["dep:rp2040-hal"]
 rp235x = ["dep:rp235x-hal"]
